refactor: rename Api to API and update imports across the codebase
