+++ src/Common/Helper.tsx
@@ -51,7 +51,7 @@ import { ReactComponent as ICRegex } from '@Icons/ic-regex.svg'
 import { ReactComponent as ICPullRequest } from '@Icons/ic-pull-request.svg'
 import { ReactComponent as ICTag } from '@Icons/ic-tag.svg'
 import { SourceTypeMap } from '@Common/Common.service'
-import { getIsRequestAborted } from './Api'
+import { getIsRequestAborted } from './API'
 
 export function showError(serverError, showToastOnUnknownError = true, hideAccessError = false) {
     if (serverError instanceof ServerErrors && Array.isArray(serverError.errors)) {
