+++ src/Common/index.ts
@@ -17,7 +17,6 @@
 export * from './Constants'
 export * from './ServerError'
 export * from './Types'
-export * from './Api'
 export { default as Reload } from './Reload'
 export { default as ErrorScreenManager } from './ErrorScreenManager'
 export { default as ErrorScreenNotAuthorized } from './ErrorScreenNotAuthorized'
@@ -66,3 +65,4 @@ export * from './SegmentedBarChart'
 export * from './CodeEditor/types'
 export * from './Tooltip'
 export * from './SegmentedControl'
+export * from './API'
