+++ src/Pages/ResourceBrowser/service.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { get, post, put, trash } from '@Common/Api'
+import { get, post, put, trash } from '@Common/API'
 import { ROUTES } from '@Common/Constants'
 import { APIOptions, ResponseType } from '@Common/Types'
 import {
